Refactor tests, add additional constraints.
