+++ model/view/soft/chart/weight.lp
@@ -1,1 +1,4 @@
-soft_weight(mark_type,c_d_temporal_agg_not_line,11).
+soft_weight(chart,c_d_temporal_agg_not_line,11).
+soft_weight(chart,only_x_over_only_y,1).
+soft_weight(chart,binned_one_dimensional,10).
+soft_weight(chart,orientation_horizontal_binned,1).
