+++ test/recommendation/cars/histogram.json
@@ -1,0 +1,21 @@
+[
+  {
+    "v1": {
+      "mark": "bar",
+      "encoding": {
+        "x": {
+          "field": "Horsepower",
+          "type": "quantitative",
+          "bin": true
+        },
+        "y": {
+          "aggregate": "count",
+          "type": "quantitative",
+          "scale": {
+            "zero": true
+          }
+        }
+      }
+    }
+  }
+]
