+++ test/recommendation/cars/histogram.lp
@@ -1,0 +1,5 @@
+view(v1).
+
+encoding(v1,e0).
+:- not field(v1,e0,"Horsepower").
+:- not bin(v1,e0,true).
