+++ test/recommendation/cars/scatter.json
@@ -1,0 +1,36 @@
+[
+  {
+    "view1": {
+      "mark": "point",
+      "encoding": {
+        "x": {
+          "field": "Horsepower",
+          "type": "quantitative",
+          "scale": { "zero": true }
+        },
+        "y": {
+          "field": "Acceleration",
+          "type": "quantitative",
+          "scale": { "zero": true }
+        }
+      }
+    }
+  },
+  {
+    "view1": {
+      "mark": "point",
+      "encoding": {
+        "x": {
+          "field": "Acceleration",
+          "type": "quantitative",
+          "scale": { "zero": true }
+        },
+        "y": {
+          "field": "Horsepower",
+          "type": "quantitative",
+          "scale": { "zero": true }
+        }
+      }
+    }
+  }
+]
