+++ test/recommendation/cars/strip.json
@@ -1,0 +1,14 @@
+[
+  {
+    "v1": {
+      "mark": "tick",
+      "encoding": {
+        "x": {
+          "field": "Horsepower",
+          "type": "quantitative",
+          "scale": { "zero": true }
+        }
+      }
+    }
+  }
+]
