+++ test/recommendation/cars/strip.lp
@@ -1,0 +1,4 @@
+view(v1).
+
+encoding(v1,e0).
+:- not field(v1,e0,"Horsepower").
