+++ test/recommendation/imdb/line.json
@@ -1,0 +1,20 @@
+[
+  {
+    "view1": {
+      "mark": "bar",
+      "encoding": {
+        "x": {
+          "type": "quantitative",
+          "aggregate": "count",
+          "scale": {
+            "zero": true
+          }
+        },
+        "y": {
+          "type": "temporal",
+          "field": "Release_Date"
+        }
+      }
+    }
+  }
+]
